[License] move cma_api to 3rdparty. separate BSD 2-clause and 3-clause
